+++ src/rules/type-rules/is-undefined.ts
@@ -13,7 +13,7 @@ export function isUndefined(options?: ValidationOptions) {
   return validator<any, unknown>(
     'isUndefined',
     (input: unknown, context: Context, _this): Nullish<any> => {
-      if (options?.coerce || context.coerce) return undefined;
+      if (options?.coerce ?? context.coerce) return undefined;
       if (input === undefined) return;
       context.fail(_this, `Value must be undefined`, input);
     },